Crate tests

metw-accounts-center tests

This crate contains test scenarios that run with either production or mock repository and client implementations. By default, cargo test uses only mock implementations. To include tests that require live services (requires a .env file):

cargo test -- --include-ignored

Some tests require human interaction and run as examples:

# Sends a verification email for adding a new address.
cargo run --example amazon-sesv2

See ACD-2 for test guideline.

Modules

account_service

Account service tests.

email_limiting_repo

Email limiting repository tests.

handlers

Hander tests.

token_repo

Token repository tests.

util

Common test utilities.