Struct collections::vec::Vec

pub struct Vec<T> {
    // some fields omitted
}

A growable list type, written Vec<T> but pronounced 'vector.'

Examples
#![feature(collections)] extern crate collections; fn main() { let mut vec = Vec::new(); vec.push(1); vec.push(2); assert_eq!(vec.len(), 2); assert_eq!(vec[0], 1); assert_eq!(vec.pop(), Some(2)); assert_eq!(vec.len(), 1); vec[0] = 7; assert_eq!(vec[0], 7); vec.push_all(&[1, 2, 3]); for x in vec.iter() { println!("{}", x); } assert_eq!(vec, [7, 1, 2, 3]); }

let mut vec = Vec::new();
vec.push(1);
vec.push(2);

assert_eq!(vec.len(), 2);
assert_eq!(vec[0], 1);

assert_eq!(vec.pop(), Some(2));
assert_eq!(vec.len(), 1);

vec[0] = 7;
assert_eq!(vec[0], 7);

vec.push_all(&[1, 2, 3]);

for x in vec.iter() {
    println!("{}", x);
}
assert_eq!(vec, [7, 1, 2, 3]);

The vec! macro is provided to make initialization more convenient:
fn main() { let mut vec = vec![1, 2, 3]; vec.push(4); assert_eq!(vec, [1, 2, 3, 4]); }

let mut vec = vec![1, 2, 3];
vec.push(4);
assert_eq!(vec, [1, 2, 3, 4]);

Use a Vec<T> as an efficient stack:
fn main() { let mut stack = Vec::new(); stack.push(1); stack.push(2); stack.push(3); while let Some(top) = stack.pop() { // Prints 3, 2, 1 println!("{}", top); } }

let mut stack = Vec::new();

stack.push(1);
stack.push(2);
stack.push(3);

while let Some(top) = stack.pop() {
    // Prints 3, 2, 1
    println!("{}", top);
}

Capacity and reallocation

The capacity of a vector is the amount of space allocated for any future elements that will be added onto the vector. This is not to be confused with the length of a vector, which specifies the number of actual elements within the vector. If a vector's length exceeds its capacity, its capacity will automatically be increased, but its elements will have to be reallocated.

For example, a vector with capacity 10 and length 0 would be an empty vector with space for 10 more elements. Pushing 10 or fewer elements onto the vector will not change its capacity or cause reallocation to occur. However, if the vector's length is increased to 11, it will have to reallocate, which can be slow. For this reason, it is recommended to use Vec::with_capacity whenever possible to specify how big the vector is expected to get.

Methods

impl<T> Vec<T>

fn new() -> Vec<T>

Constructs a new, empty Vec<T>.

The vector will not allocate until elements are pushed onto it.

Examples
fn main() { let mut vec: Vec<i32> = Vec::new(); }

let mut vec: Vec<i32> = Vec::new();

fn with_capacity(capacity: usize) -> Vec<T>

Constructs a new, empty Vec<T> with the specified capacity.

The vector will be able to hold exactly capacity elements without reallocating. If capacity is 0, the vector will not allocate.

It is important to note that this function does not specify the length of the returned vector, but only the capacity. (For an explanation of the difference between length and capacity, see the main Vec<T> docs above, 'Capacity and reallocation'.)

Examples
fn main() { let mut vec = Vec::with_capacity(10); // The vector contains no items, even though it has capacity for more assert_eq!(vec.len(), 0); // These are all done without reallocating... for i in 0..10 { vec.push(i); } // ...but this may make the vector reallocate vec.push(11); }

let mut vec = Vec::with_capacity(10);

// The vector contains no items, even though it has capacity for more
assert_eq!(vec.len(), 0);

// These are all done without reallocating...
for i in 0..10 {
    vec.push(i);
}

// ...but this may make the vector reallocate
vec.push(11);

unsafe fn from_raw_parts(ptr: *mut T, length: usize, capacity: usize) -> Vec<T>

Creates a Vec<T> directly from the raw components of another vector.

This is highly unsafe, due to the number of invariants that aren't checked.

Examples
use std::ptr; use std::mem; fn main() { let mut v = vec![1, 2, 3]; // Pull out the various important pieces of information about `v` let p = v.as_mut_ptr(); let len = v.len(); let cap = v.capacity(); unsafe { // Cast `v` into the void: no destructor run, so we are in // complete control of the allocation to which `p` points. mem::forget(v); // Overwrite memory with 4, 5, 6 for i in 0..len as isize { ptr::write(p.offset(i), 4 + i); } // Put everything back together into a Vec let rebuilt = Vec::from_raw_parts(p, len, cap); assert_eq!(rebuilt, [4, 5, 6]); } }

use std::ptr;
use std::mem;

fn main() {
    let mut v = vec![1, 2, 3];

    // Pull out the various important pieces of information about `v`
    let p = v.as_mut_ptr();
    let len = v.len();
    let cap = v.capacity();

    unsafe {
        // Cast `v` into the void: no destructor run, so we are in
        // complete control of the allocation to which `p` points.
        mem::forget(v);

        // Overwrite memory with 4, 5, 6
        for i in 0..len as isize {
            ptr::write(p.offset(i), 4 + i);
        }

        // Put everything back together into a Vec
        let rebuilt = Vec::from_raw_parts(p, len, cap);
        assert_eq!(rebuilt, [4, 5, 6]);
    }
}

unsafe fn from_raw_buf(ptr: *const T, elts: usize) -> Vec<T>

Unstable
: may be better expressed via composition

Creates a vector by copying the elements from a raw pointer.

This function will copy elts contiguous elements starting at ptr into a new allocation owned by the returned Vec<T>. The elements of the buffer are copied into the vector without cloning, as if ptr::read() were called on them.

fn capacity(&self) -> usize

Returns the number of elements the vector can hold without reallocating.

Examples
fn main() { let vec: Vec<i32> = Vec::with_capacity(10); assert_eq!(vec.capacity(), 10); }

let vec: Vec<i32> = Vec::with_capacity(10);
assert_eq!(vec.capacity(), 10);

fn reserve(&mut self, additional: usize)

Reserves capacity for at least additional more elements to be inserted in the given Vec<T>. The collection may reserve more space to avoid frequent reallocations.

Panics

Panics if the new capacity overflows usize.

Examples
fn main() { let mut vec = vec![1]; vec.reserve(10); assert!(vec.capacity() >= 11); }

let mut vec = vec![1];
vec.reserve(10);
assert!(vec.capacity() >= 11);

fn reserve_exact(&mut self, additional: usize)

Reserves the minimum capacity for exactly additional more elements to be inserted in the given Vec<T>. Does nothing if the capacity is already sufficient.

Note that the allocator may give the collection more space than it requests. Therefore capacity can not be relied upon to be precisely minimal. Prefer reserve if future insertions are expected.

Panics

Panics if the new capacity overflows usize.

Examples
fn main() { let mut vec = vec![1]; vec.reserve_exact(10); assert!(vec.capacity() >= 11); }

let mut vec = vec![1];
vec.reserve_exact(10);
assert!(vec.capacity() >= 11);

fn shrink_to_fit(&mut self)

Shrinks the capacity of the vector as much as possible.

It will drop down as close as possible to the length but the allocator may still inform the vector that there is space for a few more elements.

Examples
#![feature(collections)] extern crate collections; fn main() { let mut vec = Vec::with_capacity(10); vec.push_all(&[1, 2, 3]); assert_eq!(vec.capacity(), 10); vec.shrink_to_fit(); assert!(vec.capacity() >= 3); }

let mut vec = Vec::with_capacity(10);
vec.push_all(&[1, 2, 3]);
assert_eq!(vec.capacity(), 10);
vec.shrink_to_fit();
assert!(vec.capacity() >= 3);

fn into_boxed_slice(self) -> Box<[T]>

Converts the vector into Box<[T]>.

Note that this will drop any excess capacity. Calling this and converting back to a vector with into_vec() is equivalent to calling shrink_to_fit().

fn truncate(&mut self, len: usize)

Shorten a vector, dropping excess elements.

If len is greater than the vector's current length, this has no effect.

Examples
#![feature(collections)] extern crate collections; fn main() { let mut vec = vec![1, 2, 3, 4]; vec.truncate(2); assert_eq!(vec, [1, 2]); }

let mut vec = vec![1, 2, 3, 4];
vec.truncate(2);
assert_eq!(vec, [1, 2]);

fn as_slice(&self) -> &[T]

Unstable
: waiting on RFC revision

Extracts a slice containing the entire vector.

fn as_mut_slice(&mut self) -> &mut [T]

Unstable
: waiting on RFC revision

Deprecated: use &mut s[..] instead.

unsafe fn set_len(&mut self, len: usize)

Sets the length of a vector.

This will explicitly set the size of the vector, without actually modifying its buffers, so it is up to the caller to ensure that the vector is actually the specified size.

Examples
fn main() { let mut v = vec![1, 2, 3, 4]; unsafe { v.set_len(1); } }

let mut v = vec![1, 2, 3, 4];
unsafe {
    v.set_len(1);
}

fn swap_remove(&mut self, index: usize) -> T

Removes an element from anywhere in the vector and return it, replacing it with the last element.

This does not preserve ordering, but is O(1).

Panics

Panics if index is out of bounds.

Examples
fn main() { let mut v = vec!["foo", "bar", "baz", "qux"]; assert_eq!(v.swap_remove(1), "bar"); assert_eq!(v, ["foo", "qux", "baz"]); assert_eq!(v.swap_remove(0), "foo"); assert_eq!(v, ["baz", "qux"]); }

let mut v = vec!["foo", "bar", "baz", "qux"];

assert_eq!(v.swap_remove(1), "bar");
assert_eq!(v, ["foo", "qux", "baz"]);

assert_eq!(v.swap_remove(0), "foo");
assert_eq!(v, ["baz", "qux"]);

fn insert(&mut self, index: usize, element: T)

Inserts an element at position index within the vector, shifting all elements after position i one position to the right.

Panics

Panics if index is greater than the vector's length.

Examples
fn main() { let mut vec = vec![1, 2, 3]; vec.insert(1, 4); assert_eq!(vec, [1, 4, 2, 3]); vec.insert(4, 5); assert_eq!(vec, [1, 4, 2, 3, 5]); }

let mut vec = vec![1, 2, 3];
vec.insert(1, 4);
assert_eq!(vec, [1, 4, 2, 3]);
vec.insert(4, 5);
assert_eq!(vec, [1, 4, 2, 3, 5]);

fn remove(&mut self, index: usize) -> T

Removes and returns the element at position index within the vector, shifting all elements after position index one position to the left.

Panics

Panics if index is out of bounds.

Examples
#![feature(collections)] extern crate collections; fn main() { let mut v = vec![1, 2, 3]; assert_eq!(v.remove(1), 2); assert_eq!(v, [1, 3]); }

let mut v = vec![1, 2, 3];
assert_eq!(v.remove(1), 2);
assert_eq!(v, [1, 3]);

fn retain<F>(&mut self, f: F) where F: FnMut(&T) -> bool

Retains only the elements specified by the predicate.

In other words, remove all elements e such that f(&e) returns false. This method operates in place and preserves the order of the retained elements.

Examples
fn main() { let mut vec = vec![1, 2, 3, 4]; vec.retain(|&x| x%2 == 0); assert_eq!(vec, [2, 4]); }

let mut vec = vec![1, 2, 3, 4];
vec.retain(|&x| x%2 == 0);
assert_eq!(vec, [2, 4]);

fn push(&mut self, value: T)

Appends an element to the back of a collection.

Panics

Panics if the number of elements in the vector overflows a usize.

Examples
fn main() { let mut vec = vec!(1, 2); vec.push(3); assert_eq!(vec, [1, 2, 3]); }

let mut vec = vec!(1, 2);
vec.push(3);
assert_eq!(vec, [1, 2, 3]);

fn pop(&mut self) -> Option<T>

Removes the last element from a vector and returns it, or None if it is empty.

Examples
fn main() { let mut vec = vec![1, 2, 3]; assert_eq!(vec.pop(), Some(3)); assert_eq!(vec, [1, 2]); }

let mut vec = vec![1, 2, 3];
assert_eq!(vec.pop(), Some(3));
assert_eq!(vec, [1, 2]);

fn append(&mut self, other: &mut Self)

Unstable
: new API, waiting for dust to settle

Moves all the elements of other into Self, leaving other empty.

Panics

Panics if the number of elements in the vector overflows a usize.

Examples
#![feature(collections)] extern crate collections; fn main() { let mut vec = vec![1, 2, 3]; let mut vec2 = vec![4, 5, 6]; vec.append(&mut vec2); assert_eq!(vec, [1, 2, 3, 4, 5, 6]); assert_eq!(vec2, []); }

let mut vec = vec![1, 2, 3];
let mut vec2 = vec![4, 5, 6];
vec.append(&mut vec2);
assert_eq!(vec, [1, 2, 3, 4, 5, 6]);
assert_eq!(vec2, []);

fn drain<R>(&mut self, range: R) -> Drain<T> where R: RangeArgument<usize>

Unstable
: recently added, matches RFC

Create a draining iterator that removes the specified range in the vector and yields the removed items from start to end. The element range is removed even if the iterator is not consumed until the end.

Note: It is unspecified how many elements are removed from the vector, if the Drain value is leaked.

Panics

Panics if the starting point is greater than the end point or if the end point is greater than the length of the vector.

Examples
#![feature(collections_drain, collections_range)] extern crate collections; fn main() { // Draining using `..` clears the whole vector. let mut v = vec![1, 2, 3]; let u: Vec<_> = v.drain(..).collect(); assert_eq!(v, &[]); assert_eq!(u, &[1, 2, 3]); }

// Draining using `..` clears the whole vector.
let mut v = vec![1, 2, 3];
let u: Vec<_> = v.drain(..).collect();
assert_eq!(v, &[]);
assert_eq!(u, &[1, 2, 3]);

fn clear(&mut self)

Clears the vector, removing all values.

Examples
fn main() { let mut v = vec![1, 2, 3]; v.clear(); assert!(v.is_empty()); }

let mut v = vec![1, 2, 3];

v.clear();

assert!(v.is_empty());

fn len(&self) -> usize

Returns the number of elements in the vector.

Examples
fn main() { let a = vec![1, 2, 3]; assert_eq!(a.len(), 3); }

let a = vec![1, 2, 3];
assert_eq!(a.len(), 3);

fn is_empty(&self) -> bool

Returns true if the vector contains no elements.

Examples
fn main() { let mut v = Vec::new(); assert!(v.is_empty()); v.push(1); assert!(!v.is_empty()); }

let mut v = Vec::new();
assert!(v.is_empty());

v.push(1);
assert!(!v.is_empty());

fn map_in_place<U, F>(self, f: F) -> Vec<U> where F: FnMut(T) -> U

Unstable
: API may change to provide stronger guarantees

Converts a Vec<T> to a Vec<U> where T and U have the same size and in case they are not zero-sized the same minimal alignment.

Panics

Panics if T and U have differing sizes or are not zero-sized and have differing minimal alignments.

Examples
#![feature(collections, core)] extern crate collections; fn main() { let v = vec![0, 1, 2]; let w = v.map_in_place(|i| i + 3); assert_eq!(&w[..], &[3, 4, 5]); #[derive(PartialEq, Debug)] struct Newtype(u8); let bytes = vec![0x11, 0x22]; let newtyped_bytes = bytes.map_in_place(|x| Newtype(x)); assert_eq!(&newtyped_bytes[..], &[Newtype(0x11), Newtype(0x22)]); }

let v = vec![0, 1, 2];
let w = v.map_in_place(|i| i + 3);
assert_eq!(&w[..], &[3, 4, 5]);

#[derive(PartialEq, Debug)]
struct Newtype(u8);
let bytes = vec![0x11, 0x22];
let newtyped_bytes = bytes.map_in_place(|x| Newtype(x));
assert_eq!(&newtyped_bytes[..], &[Newtype(0x11), Newtype(0x22)]);

fn split_off(&mut self, at: usize) -> Self

Unstable
: new API, waiting for dust to settle

Splits the collection into two at the given index.

Returns a newly allocated Self. self contains elements [0, at), and the returned Self contains elements [at, len).

Note that the capacity of self does not change.

Panics

Panics if at > len.

Examples
#![feature(collections)] extern crate collections; fn main() { let mut vec = vec![1,2,3]; let vec2 = vec.split_off(1); assert_eq!(vec, [1]); assert_eq!(vec2, [2, 3]); }

let mut vec = vec![1,2,3];
let vec2 = vec.split_off(1);
assert_eq!(vec, [1]);
assert_eq!(vec2, [2, 3]);

impl<T: Clone> Vec<T>

fn resize(&mut self, new_len: usize, value: T)

Unstable
: matches collection reform specification; waiting for dust to settle

Resizes the Vec in-place so that len() is equal to new_len.

Calls either extend() or truncate() depending on whether new_len is larger than the current value of len() or not.

Examples
#![feature(collections)] extern crate collections; fn main() { let mut vec = vec!["hello"]; vec.resize(3, "world"); assert_eq!(vec, ["hello", "world", "world"]); let mut vec = vec![1, 2, 3, 4]; vec.resize(2, 0); assert_eq!(vec, [1, 2]); }

let mut vec = vec!["hello"];
vec.resize(3, "world");
assert_eq!(vec, ["hello", "world", "world"]);

let mut vec = vec![1, 2, 3, 4];
vec.resize(2, 0);
assert_eq!(vec, [1, 2]);

fn push_all(&mut self, other: &[T])

Unstable
: likely to be replaced by a more optimized extend

Appends all elements in a slice to the Vec.

Iterates over the slice other, clones each element, and then appends it to this Vec. The other vector is traversed in-order.

Examples
#![feature(collections)] extern crate collections; fn main() { let mut vec = vec![1]; vec.push_all(&[2, 3, 4]); assert_eq!(vec, [1, 2, 3, 4]); }

let mut vec = vec![1];
vec.push_all(&[2, 3, 4]);
assert_eq!(vec, [1, 2, 3, 4]);

impl<T: PartialEq> Vec<T>

fn dedup(&mut self)

Removes consecutive repeated elements in the vector.

If the vector is sorted, this removes all duplicates.

Examples
fn main() { let mut vec = vec![1, 2, 2, 3, 2]; vec.dedup(); assert_eq!(vec, [1, 2, 3, 2]); }

let mut vec = vec![1, 2, 2, 3, 2];

vec.dedup();

assert_eq!(vec, [1, 2, 3, 2]);