tests
Function get_test_file_path

pub fn get_test_file_path(path: String) -> Result<String>

Obtains a path to a test file (test code only!) This is a convenience function for writing test code. It allow tests code from anywhere in the workspace to access test files (eg input .osm files, golden outputfiles etc) which are stored within the tests package. This function make direct reference to the location of this source file (using the file!() marco) and hence should only be used in test code and not in any production code.