Crate tests

Functions

compare_with_goldenfile

Compares a string to the contents of the relevant goldenfile. Pretty prints the differences if necessary

get_test_file_path

Obtains a path to a test file (test code only!) This is a convenience function for writing test code. It allow tests code from anywhere in the workspace to access test files (eg input .osm files, golden outputfiles etc) which are stored within the tests package. This function make direct reference to the location of this source file (using the file!() marco) and hence should only be used in test code and not in any production code.

import_map

Run the contents of a .osm through the full map importer with default options.

next_test_file_path 🔒