Crate tests

Integration tests

Functions

check_proposals
Verify all edits under version control can be correctly apply to their map.

dump_route_goldenfile
Describe all public transit routes and keep under version control to spot diffs easily.

dump_turn_goldenfile
Verify what turns are generated by writing (from lane, to lane, turn type).

import_map
Run the contents of a .osm through the full map importer with default options.

main

smoke_test
Simulate an hour on every map.

test_lane_changing
Verify lane-chaging behavior is overall reasonable, by asserting all cars and bikes can complete their trip under a time limit.

test_map_importer
Test the map pipeline by importing simple, handcrafted .osm files, then emitting goldenfiles that summarize part of the generated map. Keep the goldenfiles under version control to notice when they change. The goldenfiles (and changes to them) themselves aren't easy to understand, but the test maps are.